requirements.txt
discord.py==2.3.2
requests==2.31.0
python-dotenv==1.0.0
asana==3.2.1
flask==3.0.0
gunicorn==21.2.0
apscheduler==3.10.4
sqlalchemy==2.0.23
psycopg2-binary==2.9.7
